+++ frontend/src/app.css
@@ -1,0 +1,20 @@
+:root {
+  --error-color: #f05f60;
+  --warning-color: #f0e35f;
+  --success-color: #2e7d32;
+  --default-primary-color: #751b1b;
+  --default-secondary-color: #5b1c30;
+  --white-color: #ffff;
+  --TextField-brandBorderColor: #ececec;
+  background-color: #f5f4f4;
+}
+
+.table::-webkit-scrollbar {
+  background-color: rgba(0, 0, 0, 0.1);
+  height: 7px;
+}
+
+.table::-webkit-scrollbar-thumb {
+  background-color: rgba(0, 0, 0, 0.4);
+  border-radius: 3px;
+}
