+++ frontend/src/components/headerComponent/style.css
@@ -1,0 +1,79 @@
+div .app-bar {
+  background-color: var(--white-color);
+  z-index: 99;
+  position: fixed;
+}
+
+.nav-bar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 20px;
+}
+
+.nav-bar .box-logo {
+  width: 130px;
+  height: 30px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 10px;
+  position: relative;
+}
+
+.nav-bar .box-logo img {
+  position: absolute;
+  width: 40%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.submenu {
+  width: 260px;
+  top: 64px;
+  height: 100%;
+  position: fixed;
+  background-color: var(--default-primary-color);
+  box-shadow: 0px 5px 5px 2px rgba(0, 0, 0, 0.5);
+  display: flex;
+  justify-content: center;
+}
+
+.submenu ul {
+  width: 220px;
+  display: flex;
+  justify-content: flex-start;
+  padding-top: 2px;
+  padding-left: 1px;
+  flex-direction: column;
+  height: 800px;
+  position: relative;
+  align-items: flex-start;
+}
+
+.submenu ul li {
+  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
+  color: rgb(92, 92, 92);
+  justify-content: center;
+  text-align: center;
+  display: flex;
+  padding: 20px 5px 5px 5px;
+  width: 100%;
+  font-size: 18px;
+  font-weight: 600;
+  list-style: none;
+}
+
+.submenu ul li:hover {
+  cursor: pointer;
+}
+
+.submenu ul li:active {
+  color: #666;
+  box-shadow:
+    0px 0px 0px #c5c5c5,
+    0px 0px 0px #ffffff,
+    inset 4px 4px 12px #c5c5c5,
+    inset -4px -4px 12px #ffffff;
+    border-radius: 10px;
+}
