+++ frontend/src/pages/home/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Header from '../../components/headerComponent';
+
+
+const Home = () => {
+
+  return (
+    <div>
+      <Header />
+    </div>
+  )
+}
+
+export default Home;
