+++ frontend/src/pages/login/index.jsx
@@ -7,6 +7,9 @@ const Login = () => {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
 
+
+
+
   return (
     <div className="background-login">
       <div class="form-control">
@@ -41,7 +44,6 @@ const Login = () => {
           <div className="box-submit-login">
             <button class="submit-btn">Sign In</button>
           </div>
-          
         </form>
       </div>
     </div>
