--- frontend/src/routes/index.js
@@ -1,16 +1,0 @@
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import Login from '../pages/login'
-
-
-const Routers = () => {
-
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/login" element={<Login />} />
-      </Routes>
-    </BrowserRouter>
-  )
-};
-
-export default Routers;
