+++ frontend/src/routes/index.jsx
@@ -1,0 +1,16 @@
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Login from "../pages/login";
+import Home from "../pages/home";
+
+const Routers = () => {
+  return (
+    <BrowserRouter>
+      <Routes>
+        <Route path="/login" element={<Login />} />
+        <Route path="/home" element={<Home />} />
+      </Routes>
+    </BrowserRouter>
+  );
+};
+
+export default Routers;
